Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
Pillow==10.1.0
cloudinary==1.36.0
dj-database-url==2.1.0
django-cloudinary-storage==0.3.0
django-filter==23.5